feat: 네이버 채용 로직 개선
